pyproject.toml
[tool.poetry]
name = "cursovaya-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-stubs = "^5.0.4"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.27.0"
redis = "^5.0.8"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
coverage = "^7.6.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
